+++ package.json
@@ -62,6 +62,9 @@
     ]
   },
   "pnpm": {
+    "overrides": {
+      "@babel/types": "7.24.0"
+    },
     "patchedDependencies": {
       "@stitches/react@1.3.1-1": "patches/@stitches__react@1.3.1-1.patch",
       "css-tree@2.3.1": "patches/css-tree@2.3.1.patch",
